Extract ParCount to a separate file
