+++ rholang/src/main/scala/coop/rchain/rholang/interpreter/matcher/SpatialMatcher.scala
@@ -397,107 +397,6 @@ object SpatialMatcher extends SpatialMatcherInstances {
       _                   <- freeMap.modify(_ + (level -> remainderParUpdated))
     } yield Unit
 
-  final case class ParCount(
-      sends: Int = 0,
-      receives: Int = 0,
-      news: Int = 0,
-      exprs: Int = 0,
-      matches: Int = 0,
-      ids: Int = 0,
-      bundles: Int = 0
-  ) {
-    def min(other: ParCount): ParCount = binOp(math.min, other)
-
-    def max(other: ParCount): ParCount = binOp(math.max, other)
-
-    def +(other: ParCount): ParCount = binOp(saturatingAdd, other)
-
-    def binOp(op: (Int, Int) => Int, other: ParCount): ParCount =
-      ParCount(
-        sends = op(sends, other.sends),
-        receives = op(receives, other.receives),
-        news = op(news, other.news),
-        exprs = op(exprs, other.exprs),
-        matches = op(matches, other.matches),
-        ids = op(ids, other.ids),
-        bundles = op(bundles, other.bundles)
-      )
-
-    // Only saturates going from positive to negative
-    def saturatingAdd(l: Int, r: Int): Int = {
-      val res = l + r
-      (res | -(if (res < l) 1 else 0)) & ~Int.MinValue
-    }
-  }
-
-  object ParCount {
-    def apply(par: Par): ParCount =
-      ParCount(
-        sends = par.sends.size,
-        receives = par.receives.size,
-        news = par.news.size,
-        matches = par.matches.size,
-        exprs = par.exprs.size,
-        ids = par.ids.size,
-        bundles = par.bundles.size
-      )
-    def max: ParCount =
-      ParCount(
-        sends = Int.MaxValue,
-        receives = Int.MaxValue,
-        news = Int.MaxValue,
-        matches = Int.MaxValue,
-        exprs = Int.MaxValue,
-        ids = Int.MaxValue,
-        bundles = Int.MaxValue
-      )
-
-    def minMax(par: Par): (ParCount, ParCount) = {
-      val pc = ParCount(noFrees(par))
-      val wildcard: Boolean = par.exprs.exists { expr =>
-        expr.exprInstance match {
-          case EVarBody(EVar(v)) =>
-            v.varInstance match {
-              case Wildcard(_) => true
-              case FreeVar(_)  => true
-              case _           => false
-            }
-          case _ => false
-        }
-      }
-      val minInit = pc
-      val maxInit = if (wildcard) ParCount.max else pc
-
-      par.connectives.foldLeft((minInit, maxInit)) {
-        case ((min, max), con) =>
-          val (cmin, cmax) = minMax(con)
-          (min + cmin, max + cmax)
-      }
-    }
-
-    def minMax(con: Connective): (ParCount, ParCount) =
-      con.connectiveInstance match {
-        case ConnAndBody(ConnectiveBody(ps)) =>
-          val pMinMax = ps.map(minMax)
-          (pMinMax.foldLeft(ParCount())(_ max _._1), pMinMax.foldLeft(ParCount.max)(_ min _._2))
-        case ConnOrBody(ConnectiveBody(ps)) =>
-          val pMinMax = ps.map(minMax)
-          (pMinMax.foldLeft(ParCount.max)(_ min _._1), pMinMax.foldLeft(ParCount())(_ max _._2))
-        case ConnNotBody(_) =>
-          (ParCount(), ParCount.max)
-        case ConnectiveInstance.Empty =>
-          (ParCount(), ParCount())
-        case _: VarRefBody =>
-          // Variable references should be substituted before going into the matcher.
-          // This should never happen.
-          (ParCount(), ParCount())
-        case _: ConnBool      => (ParCount(exprs = 1), ParCount(exprs = 1))
-        case _: ConnInt       => (ParCount(exprs = 1), ParCount(exprs = 1))
-        case _: ConnString    => (ParCount(exprs = 1), ParCount(exprs = 1))
-        case _: ConnUri       => (ParCount(exprs = 1), ParCount(exprs = 1))
-        case _: ConnByteArray => (ParCount(exprs = 1), ParCount(exprs = 1))
-      }
-  }
 }
 
 trait SpatialMatcherInstances {
